Add and apply .editorconfig from P2 (#1195)
* Consistenty use javax annotations.
 - Unfortunately jetbrains annotations seem to be exposed transitively via core somewhere, but with the correct IDE settings, annotations can be defaulted to javax
 - Cleaning up of import order in #1195
 - Must be merged before #1195

* Add and apply .editorconfig from P2
 - Does not rearrange entries

* Address some comments

* add back some javadoc comments

* Address final comments

+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/math/MutableVector3.java
@@ -19,7 +19,7 @@ public class MutableVector3 extends Vector3 {
     }
 
     public MutableVector3(float x, float y, float z) {
-        this((double) x, (double) y, (double) z);
+        this(x, y, (double) z);
     }
 
     public MutableVector3(Vector3 other) {
